ElevenLabs/backend/classifier_mediapipe.py
"""
Sound event classification using MediaPipe Audio Classifier
TODO: Full MediaPipe integration requires model asset setup
Currently provides placeholder implementation
"""

import logging
import numpy as np
from typing import Optional
from audio_stream import AudioStream

logger = logging.getLogger(__name__)


class MediaPipeClassifier:
    """
    Sound event classifier using MediaPipe
    TODO: Requires MediaPipe Audio Classifier model assets
    See: https://ai.google.dev/edge/mediapipe/solutions/audio/audio_classifier
    
    For now, provides placeholder mapping based on energy levels
    """
    
    def __init__(self):
        self.initialized = False
        
    def initialize(self):
        """Initialize MediaPipe classifier (placeholder)"""
        # TODO: Initialize MediaPipe Audio Classifier
        # Example:
        # import mediapipe as mp
        # self.mp_audio = mp.solutions.audio
        # self.classifier = self.mp_audio.AudioClassifier(
        #     model_path="path/to/model.tflite",
        #     num_threads=4
        # )
        
        logger.warning("MediaPipe classifier not fully implemented - using placeholder")
        self.initialized = True
    
    def classify(self, audio: np.ndarray, sample_rate: int = 16000) -> str:
        """
        Classify sound event in audio chunk
        
        Args:
            audio: Audio samples as numpy array (float32, mono)
            sample_rate: Sample rate in Hz (default 16000)
        
        Returns:
            Sound event label string (e.g., "[DOG BARK]", "[SIREN]", etc.)
        """
        if not self.initialized:
            self.initialize()
        
        # TODO: Replace with actual MediaPipe classification
        # Example:
        # results = self.classifier.classify(audio, sample_rate)
        # top_result = results.classifications[0].categories[0]
        # return f"[{top_result.category_name}]"
        
        # Placeholder: Simple energy-based mapping
        energy = AudioStream.get_rms_energy(audio)
        
        if energy > 0.15:
            return "[LOUD_NOISE]"
        elif energy > 0.08:
            return "[MODERATE_NOISE]"
        elif energy > 0.04:
            return "[QUIET_NOISE]"
        else:
            return "[SILENCE]"
    
    @staticmethod
    def get_placeholder_labels() -> list[str]:
        """Return list of placeholder labels for testing"""
        return [
            "[LOUD_NOISE]",
            "[MODERATE_NOISE]",
            "[QUIET_NOISE]",
            "[SILENCE]",
            "[DOG BARK]",
            "[SIREN]",
            "[DOOR SLAM]",
            "[GLASS BREAK]"
        ]


# TODO: MediaPipe Audio Classifier Setup Instructions
"""
To fully implement MediaPipe Audio Classifier:

1. Install MediaPipe:
   pip install mediapipe

2. Download Audio Classifier model:
   - Visit: https://ai.google.dev/edge/mediapipe/solutions/audio/audio_classifier
   - Download the YAMNet model or custom model TFLite file
   - Place in models/ directory

3. Update initialize() method:
   import mediapipe as mp
   mp_audio = mp.solutions.audio
   self.classifier = mp_audio.AudioClassifier(
       model_path="models/yamnet.tflite",
       num_threads=4
   )

4. Update classify() method to use actual classifier results

5. Handle model-specific input requirements (sample rate, chunk size, etc.)
"""
