ElevenLabs/requirements.txt
# SoundSight Backend Dependencies

# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# Serial communication
pyserial>=3.5

# Speech-to-Text
requests>=2.31.0

# WebSocket server
websockets>=12.0

# Optional: MediaPipe for sound classification
# mediapipe>=0.10.0

# Logging (built-in, but can use structlog for better formatting)
# structlog>=23.0.0
